requirements.txt
Django
djangorestframework
mysql
mysql-connector
mysqlclient
pandas
PyMySQL
SQLAlchemy==1.4.46
